MATLAB Surface Plot
A script to plot the 3D bandstructure of a material. The script reads in the (x,y) coordinates in k-space and the associated eigenvalues. This is then repeated to create a supercell view of the 3D bands. A hexagon is then overlaid with the plot to represent the edge of the first Brillouin Zone.